Use liquid lang and added custom tooltips
